routes: rename `server` module to `routes`
