+++ src/lib.rs
@@ -1,4 +1,4 @@
 pub mod configuration;
 pub mod domain;
-pub mod server;
+pub mod routes;
 pub mod start_up;
